+++ packages/transaction-pay-controller/src/utils/transaction.test.ts
@@ -8,7 +8,7 @@ import {
 import type { TransactionMeta } from '@metamask/transaction-controller';
 import type { TransactionControllerState } from '@metamask/transaction-controller';
 import type { Hex } from '@metamask/utils';
-import { noop } from 'lodash';
+import { noop } from 'lodash-es';
 
 import { NATIVE_TOKEN_ADDRESS } from '../constants.js';
 import { getMessengerMock } from '../tests/messenger-mock.js';